ai_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
pandas
numpy
faiss-cpu
sentence-transformers
openai
